refactor: repackaging output
