+++ cmd/ocm-addons/notify/list/cmd.go
@@ -6,7 +6,6 @@ import (
 
 	"github.com/mt-sre/ocm-addons/internal/cli"
 	"github.com/mt-sre/ocm-addons/internal/notification"
-	"github.com/mt-sre/ocm-addons/internal/output"
 
 	"github.com/apex/log"
 	"github.com/spf13/cobra"
@@ -52,11 +51,11 @@ func run(opts *options) func(*cobra.Command, []string) error {
 
 		defer sess.End()
 
-		table, err := output.NewTable(
-			output.WithColumns(opts.Columns),
-			output.WithNoColor(opts.NoColor),
-			output.WithNoHeaders(opts.NoHeaders),
-			output.WithPager(sess.Pager()),
+		table, err := cli.NewTable(
+			cli.WithColumns(opts.Columns),
+			cli.WithNoColor(opts.NoColor),
+			cli.WithNoHeaders(opts.NoHeaders),
+			cli.WithPager(sess.Pager()),
 		)
 		if err != nil {
 			return fmt.Errorf("creating table: %w", err)
@@ -76,7 +75,7 @@ func run(opts *options) func(*cobra.Command, []string) error {
 				for id, cfg := range configs {
 					cfg := cfg
 
-					if err := table.Write(&cfg, output.WithAdditionalFields(
+					if err := table.Write(&cfg, cli.WithAdditionalFields(
 						map[string]interface{}{
 							"ID":      id,
 							"Product": prod,
