+++ internal/ocm/addon_installation_test.go
@@ -1,12 +1,13 @@
-package ocm
+package ocm_test
 
 import (
 	"testing"
 
-	"github.com/mt-sre/ocm-addons/internal/output"
+	"github.com/mt-sre/ocm-addons/internal/cli"
+	"github.com/mt-sre/ocm-addons/internal/ocm"
 	"github.com/stretchr/testify/require"
 )
 
 func TestAddonInstallationInterfaces(t *testing.T) {
-	require.Implements(t, new(output.RowDataProvider), new(AddonInstallation))
+	require.Implements(t, new(cli.RowDataProvider), new(ocm.AddonInstallation))
 }
