+++ internal/ocm/cluster_test.go
@@ -1,16 +1,17 @@
-package ocm
+package ocm_test
 
 import (
 	"testing"
 
-	"github.com/mt-sre/ocm-addons/internal/output"
+	"github.com/mt-sre/ocm-addons/internal/cli"
+	"github.com/mt-sre/ocm-addons/internal/ocm"
 	"github.com/stretchr/testify/require"
 )
 
 func TestClusterInterfaces(t *testing.T) {
-	require.Implements(t, new(output.RowDataProvider), new(Cluster))
+	require.Implements(t, new(cli.RowDataProvider), new(ocm.Cluster))
 }
 
 func TestSubscriptionInterfaces(t *testing.T) {
-	require.Implements(t, new(output.RowDataProvider), new(Subscription))
+	require.Implements(t, new(cli.RowDataProvider), new(ocm.Subscription))
 }
